Major rehaul
Major change to allow training with non-matching sequence lengths. It was found that the jogging class had shorter temporal resolution, which made the network believe that any padded frames were jogging (as it had to be padded the most).

Next fix has to be to allow lazy loading (cache preprocessing for various datapoints to allow bucketized batching.